requirements/_base.txt
-r _base.txt
django-debug-toolbar==3.2.1
Faker==8.1.2
## The following requirements were added by pip freeze:
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.3
django-filter==2.4.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-rw-serializers==1.0.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
openapi-codec==1.3.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
